feat: add README
