+++ src/components/Button.tsx
@@ -1,20 +1,20 @@
-export const Button = ({
-  changeSortParams,
-  sortParam,
-}: {
-  changeSortParams: (param: string) => void;
-  sortParam: string;
-}) => {
-  const onClick = (event: React.MouseEvent<HTMLInputElement>) => {
-    // console.log(event.currentTarget.value)
-    changeSortParams(event.currentTarget.value);
-  };
-  return (
-    <input
-      type="button"
-      className="rounded w-30 h-7 bg-amber-500"
-      value={sortParam}
-      onClick={onClick}
-    ></input>
-  );
-};
+import { memo } from 'react';
+
+export const Button = memo(
+  ({
+    changeSortParams,
+    sortParam,
+  }: {
+    changeSortParams: (event: React.MouseEvent<HTMLInputElement>) => void;
+    sortParam: string;
+  }) => {
+    return (
+      <input
+        type="button"
+        className="rounded w-30 h-7 bg-amber-500"
+        value={sortParam}
+        onClick={changeSortParams}
+      ></input>
+    );
+  }
+);
