+++ src/components/SearchInput .tsx
@@ -1,15 +1,17 @@
-import type { ChangeEvent } from 'react';
+import { memo, type ChangeEvent } from 'react';
 
-export const SearchInput = ({
-  changeSearchInput,
-}: {
-  changeSearchInput: (event: ChangeEvent<HTMLInputElement>) => void;
-}) => {
-  return (
-    <input
-      type="text"
-      className="border rounded-2xl pl-5"
-      onChange={changeSearchInput}
-    />
-  );
-};
+export const SearchInput = memo(
+  ({
+    changeSearchInput,
+  }: {
+    changeSearchInput: (event: ChangeEvent<HTMLInputElement>) => void;
+  }) => {
+    return (
+      <input
+        type="text"
+        className="border rounded-2xl pl-5"
+        onChange={changeSearchInput}
+      />
+    );
+  }
+);
