+++ src/components/SortSelector.tsx
@@ -1,25 +1,26 @@
+import { memo } from 'react';
 import { Button } from './Button';
 
-export const SortSelector = ({
-  changeSortParams,
-  sortParams,
-}: {
-  changeSortParams: (param: string) => void;
-  sortParams: string[];
-}) => {
-  // const
-
-  return (
-    <div className="flex gap-1">
-      {sortParams.map((param) => {
-        return (
-          <Button
-            key={param}
-            changeSortParams={changeSortParams}
-            sortParam={param}
-          />
-        );
-      })}
-    </div>
-  );
-};
+export const SortSelector = memo(
+  ({
+    changeSortParams,
+    sortParams,
+  }: {
+    changeSortParams: (event: React.MouseEvent<HTMLInputElement>) => void;
+    sortParams: string[];
+  }) => {
+    return (
+      <div className="flex gap-1">
+        {sortParams.map((param) => {
+          return (
+            <Button
+              key={param}
+              changeSortParams={changeSortParams}
+              sortParam={param}
+            />
+          );
+        })}
+      </div>
+    );
+  }
+);
